Authorization bug - bookmark update and delete buttons not showing (checked in for review)

+++ app/controllers/bookmarks_controller.rb
@@ -1,8 +1,15 @@
 class BookmarksController < ApplicationController
 
+  def new
+    @topic = Topic.find(params[:topic_id])
+    @bookmark = Bookmark.new
+    authorize @bookmark
+  end
 
   def show
+    @topic = Topic.find(params[:topic_id])
     @bookmark = Bookmark.find(params[:id])
+    authorize @bookmark
   end
 
   def edit
@@ -10,19 +17,11 @@ class BookmarksController < ApplicationController
     @bookmark = Bookmark.find(params[:id])
   end
 
-  def update
-    @topic = Topic.find(params[:topic_id])
-    @bookmark = Bookmark.find(params[:id])
-
-    @bookmark.update_attributes(params.require(:bookmark).permit(:url))
-
-    redirect_to @topic
-  end
-
   def create
     @topic = Topic.find(params[:topic_id])
     @bookmark = Bookmark.new(params.require(:bookmark).permit(:url))
     @bookmark.topic = @topic
+    authorize @bookmark
 
     if @bookmark.save
       flash[:notice] = "\"#{@bookmark.url}\" was saved."
@@ -33,10 +32,24 @@ class BookmarksController < ApplicationController
     end
   end
 
+  def update
+    @topic = Topic.find(params[:topic_id])
+    @bookmark = Bookmark.find(params[:id])
+    authorize @bookmark
+    if @bookmark.update_attributes(params.require(:bookmark).permit(:url))
+      flash[:notice] = "Bookmark was updated"
+      redirect_to @topic
+    else
+      flash[:error] = "The bookmark could not be updated. Please try again."
+      render :edit
+    end
+
+  end
+
   def destroy
     @topic = Topic.find(params[:topic_id])
     @bookmark = Bookmark.find(params[:id])
-
+    authorize @bookmark
     if @bookmark.destroy
       flash[:notice] = "Bookmark was deleted"
     else
